fix more test issues
