+++ droidlet/dialog/post_process_logical_form.py
@@ -67,3 +67,22 @@ def fix_reference_object_with_filters(d):
             new_d[key] = fix_reference_object_with_filters(value)
 
     return new_d
+
+
+def retrieve_ref_obj_span(d: dict):
+    if d.get("pred_text") == "has_name":
+        return d["obj_text"]
+
+    for k, v in d.items():
+        if isinstance(v, dict):
+            obj = retrieve_ref_obj_span(v)
+            if obj:
+                return obj
+        elif isinstance(v, list):
+            for item in v:
+                if isinstance(item, dict):
+                    obj = retrieve_ref_obj_span(item)
+                    if obj:
+                        return obj
+
+    return None
